apriori-update-lambda/requirements.txt
boto3
pandas<2.0.0
psycopg2-binary
apyori